+++ lib/dio/public_api.rb
@@ -1,4 +1,8 @@
 module Dio
+  # Public API for Dio
+  #
+  # @author [baweaver]
+  # @since 0.0.1
   module PublicApi
     # Treats `[]` like an alternative constructor and forwards to `DiveForwarder`
     #
@@ -7,6 +11,31 @@ module Dio
     #
     # @return [Dio::DiveForwarder]
     #   Dio pattern matching interface
-    def [](...) = Dio::DiveForwarder.new(...)
+    def [](...) = Dio::Forwarders::BaseForwarder.new(...)
+
+    # Dynamic Forwarder, uses `public_send` for Hash forwarding
+    #
+    # @param ... [Any]
+    #   Arguments to match against
+    #
+    # @return [Dio::Forwarders::BaseForwarder]
+    def dynamic(...) = Dio::Forwarders::BaseForwarder.new(...)
+
+    # Attribute Forwarder, extracts `attr_*` methods to match against
+    #
+    # @param ... [Any]
+    #   Arguments to match against
+    #
+    # @return [Dio::Forwarders::AttributeForwarder]
+    def attribute(...) = Dio::Forwarders::AttributeForwarder.new(...)
+
+    # String Hash Forwarder, treats a String Hash like a Symbol Hash for
+    # matching against.
+    #
+    # @param ... [Any]
+    #   Arguments to match against
+    #
+    # @return [Dio::Forwarders::StringHashForwarder]
+    def string_hash(...) = Dio::Forwarders::StringHashForwarder.new(...)
   end
 end