+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## Unreleased
+
+* None
+
+## 0.0.3 - 01/17/2021
+
+* Forwarders introduced
+  * Base concept of Forwarder and two new Forwarders introduced
+  * Attribute Forwarder to only work on `attr_*` methods, narrowing scope
+  * String Hash Forwarder to work on String keyed Hashes
+
 ## 0.0.2 - 01/17/2021
 
 * [Patch 1](https://github.com/baweaver/dio/issues/1) - Lock Ruby version to 3.x